django-celery/project/requirements.txt
Django
pytest
pytest-django

celery
redis




opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-celery
opentelemetry-instrumentation-django
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-asgi
opentelemetry-sdk
gunicorn
requests
uvicorn[standard]
